requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.6.0
click==8.1.3
Django==4.1
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
promise==2.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
